% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{coef.pliable}
\alias{coef.pliable}
\title{Return estimated coefficients from a fitted pliable model.}
\usage{
\method{coef}{pliable}(object, lambda = NULL, ...)
}
\arguments{
\item{object}{object returned from a call to pliable}

\item{lambda}{values of lambda at whcih predictions are desired. If NULL (default), the path of lambda values from the fitted model.
are used. If lambda is  not NULL,  the predictions are made at the closest values to lambda in the lambda path  from  the fitted model;}

\item{...}{Further arguments (not used)

 @return  estimated parameters}
}
\description{
Return estimated coefficients from a fitted pliable model.
}
